requirements (5).txt
gradio>=3.0
torch>=1.10
transformers>=4.20
numpy>=1.21
matplotlib>=3.5
pillow>=9.0
opencv-python>=4.5
scikit-image>=0.19